requirements/dev.txt
slack_sdk
python-docx
spacy
presidio-anonymizer==2.2.355
presidio-analyzer==2.2.355
Faker==28.0.0
fuzzywuzzy==0.18.0
python-Levenshtein
openpyxl 
pandas==2.2.3
olefile
beautifulsoup4
pydub==0.25.1
tabulate
flask-mail
werkzeug
pyarrow
jsonschema
scikit-learn
pyTigerGraph==1.8.6
scrapy
xlrd
smart_open[boto3]
xgboost==1.7.6
shap==0.44.0
docx2txt
ddgs
pypandoc==1.16.2